suicide (almost c'est faux)

+++ grayscale.c
@@ -17,7 +17,7 @@ char* Grayscale(char path[])
 	{
 		for(int y = 0; y < img->h; y++)
 		{
-			Uint8 pixel = GetPixel(img, x, y);
+			Uint32 pixel = GetPixel(img, x, y);
 			SDL_GetRGB(pixel, img->format, &r, &g, &b);
 			Uint8 gray = (Uint8)(r * 0.3f + g * 0.59f + b * 0.11f);
 			Uint32 averagePixel = SDL_MapRGB(img->format, gray, gray, gray);
